Moved everything in components to single-product-backup
