+++ _inc/client/plans/single-product-backup.jsx
@@ -8,7 +8,6 @@ import { translate as __ } from 'i18n-calypso';
  * Internal dependencies
  */
 import { AccentedCard } from './accented-card';
-import { PlanPriceDisplay, PlanRadioButton } from './components';
 import Gridicon from 'components/gridicon';
 
 export function SingleProductBackup( props ) {
@@ -57,6 +56,37 @@ function SingleProductBackupHeader( props ) {
 					yearlyPrice={ sitePlans[ 'realtime-backup' ].price.yearly.amount }
 				/>
 			) }
+		</div>
+	);
+}
+
+export function PlanPriceDisplay( props ) {
+	const { dailyPrice, yearlyPrice } = props;
+	const perYearPriceRange = `${ dailyPrice }-${ yearlyPrice } /year`;
+
+	return (
+		<div
+			style={ {
+				display: 'flex',
+				flexDirection: 'row',
+				justifyContent: 'space-between',
+				alignContent: 'center',
+			} }
+		>
+			<SlashedPrice />
+			<div className="plans-price__container">
+				<span className="plans-price__span">{ perYearPriceRange }</span>
+			</div>
+		</div>
+	);
+}
+
+function SlashedPrice() {
+	return (
+		<div className="slashed-price__container" style={ { marginRight: '14px' } }>
+			<div className="slashed-price__slash"></div>
+			{ /* TODO: get this from an API or calculate, currently unsure how to get this increased price */ }
+			<div className="slashed-price__price">{ '$15-25' }</div>
 		</div>
 	);
 }
@@ -126,6 +156,35 @@ function SingleProductBackupBody( props ) {
 				<a href={ upgradeLinks[ selectedBackupType ] } type="button" class="dops-button is-primary">
 					{ upgradeTitles[ selectedBackupType ] }
 				</a>
+			</div>
+		</div>
+	);
+}
+
+export function PlanRadioButton( props ) {
+	const { checked, onChange, planName, radioValue, planPrice } = props;
+
+	return (
+		<div className="plans-section__radio-toggle">
+			<div
+				style={ {
+					display: 'flex',
+					justifyContent: 'center',
+					alignItems: 'center',
+				} }
+			>
+				<input
+					style={ { gridColumn: 1, gridRow: 1 } }
+					type="radio"
+					value={ radioValue }
+					checked={ checked }
+					onChange={ onChange }
+				/>
+			</div>
+			<div style={ { gridColumn: 2, gridRow: 1, fontWeight: 'bold' } }>{ planName }</div>
+			<div style={ { gridColumn: 2, gridRow: 2 } }>
+				{ /* TODO: how to I18N this? */ }
+				{ planPrice && `${ planPrice } /year` }
 			</div>
 		</div>
 	);
